+++ src/trees.rs
@@ -1,7 +1,10 @@
 use imgui_sys;
 use std::marker::PhantomData;
 
-use super::{ImGuiSetCond, Ui};
+use super::{ImGuiSetCond, ImGuiTreeNodeFlags, ImGuiTreeNodeFlags_Bullet,
+            ImGuiTreeNodeFlags_DefaultOpen, ImGuiTreeNodeFlags_Leaf,
+            ImGuiTreeNodeFlags_OpenOnArrow, ImGuiTreeNodeFlags_OpenOnDoubleClick,
+            ImGuiTreeNodeFlags_Selected, Ui};
 
 #[must_use]
 pub struct TreeNode<'ui, 'p> {
@@ -35,7 +38,7 @@ impl<'ui, 'p> TreeNode<'ui, 'p> {
     pub fn build<F: FnOnce()>(self, f: F) {
         let render = unsafe {
             if !self.opened_cond.is_empty() {
-                imgui_sys::igSetNextTreeNodeOpened(self.opened, self.opened_cond);
+                imgui_sys::igSetNextTreeNodeOpen(self.opened, self.opened_cond);
             }
             match self.label {
                 Some(label) => {
@@ -49,5 +52,58 @@ impl<'ui, 'p> TreeNode<'ui, 'p> {
             f();
             unsafe { imgui_sys::igTreePop() };
         }
+    }
+}
+
+#[must_use]
+pub struct CollapsingHeader<'ui, 'p> {
+    label: &'p str,
+    // Some flags are automatically set in ImGui::CollapsingHeader, so
+    // we only support a sensible subset here
+    flags: ImGuiTreeNodeFlags,
+    _phantom: PhantomData<&'ui Ui<'ui>>,
+}
+
+impl<'ui, 'p> CollapsingHeader<'ui, 'p> {
+    pub fn new(label: &'p str) -> Self {
+        CollapsingHeader {
+            label: label,
+            flags: ImGuiTreeNodeFlags::empty(),
+            _phantom: PhantomData,
+        }
+    }
+    #[inline]
+    pub fn flags(self, flags: ImGuiTreeNodeFlags) -> Self {
+        CollapsingHeader { flags: flags, ..self }
+    }
+    #[inline]
+    pub fn selected(self, value: bool) -> Self {
+        CollapsingHeader { flags: self.flags.with(ImGuiTreeNodeFlags_Selected, value), ..self }
+    }
+    #[inline]
+    pub fn default_open(self, value: bool) -> Self {
+        CollapsingHeader { flags: self.flags.with(ImGuiTreeNodeFlags_DefaultOpen, value), ..self }
+    }
+    #[inline]
+    pub fn open_on_double_click(self, value: bool) -> Self {
+        CollapsingHeader {
+            flags: self.flags.with(ImGuiTreeNodeFlags_OpenOnDoubleClick, value),
+            ..self
+        }
+    }
+    #[inline]
+    pub fn open_on_arrow(self, value: bool) -> Self {
+        CollapsingHeader { flags: self.flags.with(ImGuiTreeNodeFlags_OpenOnArrow, value), ..self }
+    }
+    #[inline]
+    pub fn leaf(self, value: bool) -> Self {
+        CollapsingHeader { flags: self.flags.with(ImGuiTreeNodeFlags_Leaf, value), ..self }
+    }
+    #[inline]
+    pub fn bullet(self, value: bool) -> Self {
+        CollapsingHeader { flags: self.flags.with(ImGuiTreeNodeFlags_Bullet, value), ..self }
+    }
+    pub fn build(self) -> bool {
+        unsafe { imgui_sys::igCollapsingHeader(imgui_sys::ImStr::from(self.label), self.flags) }
     }
 }